requirements.txt
aiofiles==24.1.0
aiohttp==3.11.16
discord.py==2.6.3
python-dotenv==1.1.1
